+++ docs/changelog.rst
@@ -2,6 +2,11 @@ v6.1.0 (To be release)
 =========================
 Feat: Added cloud_connector API Talker for Cloud and Branch Connector (by `Dominic Schimanski <mailto:dschimanski@zscaler.com>)
 Feat: Added Method to update APP connectors (by `Sergio Pereira <mailto:spereira@zscaler.com>`_)
+FIX: Proper defaults for ZIA __init__() attributes.  (by `Dax Mickelson <mailto:dmickelson@zscaler.com>`_)
+FIX: All instead of Any for ZIA username/password/api_key auth method.  (by `Dax Mickelson <mailto:dmickelson@zscaler
+.com>`_)
+Feat: Add warning message about needing to auth if not done during ZIA __init__().  (by `Dax Mickelson
+<mailto:dmickelson@zscaler.com>`_)
 
 v6.0.0 (August 2023)
 =========================